Committed changes as per discussion
