+++ src/main/resources/config.properties
@@ -18,9 +18,6 @@ ms_password=
 native_username=rramos@deloitte.com
 native_password=Test123!
 
-author_user=
-author_password=
-
 ## Test data file paths
 BIDataImportCSVFile=./src/test/resources/data/BIDataImportsTestData.csv
 BIDataImportExcelFile=./src/test/resources/data/BIDataImportsTestData.xlsx
