+++ src/test/resources/Features/Login.feature
@@ -4,3 +4,9 @@ Feature: AI Core application
     Given User is on application
     When User enters username and password and click on SignIn button
     Then User can navigate to home page
+    
+  Scenario: Login with native credentials
+    Given User is on application
+    When User enters nativeUsername and nativePassword
+    And User clicks on Login with native button
+    Then User can navigate to home page
